refactor: Merge utils.search_utils into utils.common
